Isolated configuration data
Note that config data gets dumped into the top-level of the data structure in
project.clj, as such we were littering that top-level with all of our values.
With this change, the only value that gets dumped in there is the `:env` key
and associated data. As such, the bits that read the configuration (the only
code that does that is now limited to the components) needed to be updated for
this additional level of structure.